Move LearnImmersiveLayout out of LearnIndex and use same router pattern
